Split SortedCollection_spec.js into one file per method
- SortedCollection_constructor_spec.js
- SortedCollection_push_spec.js
- SortedCollection_list_spec.js
- SortedCollection_size_spec.js
- SortedCollection_select_spec.js
- SortedCollection_after_spec.js
- SortedCollection_from_spec.js
- SortedCollection_before_spec.js
- SortedCollection_upTo_spec.js
- SortedCollection_hasAny_spec.js
